requirements.txt
 
crewai
crewai_tools
python-dotenv
langchain-openai
langchain-groq
google-search-results
langchain-google-genai